Set bigIncrements instead of Increments

+++ database/migrations/2019_12_11_101416_create_threads_table.php
@@ -14,8 +14,8 @@ class CreateThreadsTable extends Migration
     public function up()
     {
         Schema::create('threads', function (Blueprint $table) {
-            $table->increments('id');
-            $table->unsignedInteger('user_id');
+            $table->bigIncrements('id');
+            $table->unsignedBigInteger('user_id');
             $table->string('slug')->index();
             $table->string('title');
             $table->text('body');
